完成SpringSecurity整合，添加日志功能，完善代码格式

+++ src/main/java/com/relaper/cr/springrbac/service/impl/MenuServiceImpl.java
@@ -1,6 +1,7 @@
 package com.relaper.cr.springrbac.service.impl;
 
 import com.relaper.cr.springrbac.dao.MenuDao;
+import com.relaper.cr.springrbac.dao.RoleMenuDao;
 import com.relaper.cr.springrbac.dto.MenuDto;
 import com.relaper.cr.springrbac.dto.MenuIndexDto;
 import com.relaper.cr.springrbac.entity.MyMenu;
@@ -20,9 +21,11 @@ import java.util.List;
 public class MenuServiceImpl implements MenuService {
     @Autowired
     private MenuDao menuDao;
+    @Autowired
+    private RoleMenuDao roleMenuDao;
+
     @Override
     public List<MyMenu> getMenuAll(String queryName,Integer queryType) {
-
         return menuDao.getFuzzyMenu(queryName,queryType);
     }
 
@@ -33,28 +36,44 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     public List<MenuDto> buildMenuAll() {
-
         return menuDao.buildAll();
     }
 
     @Override
     public Result updateMenu(MyMenu menu) {
+        menu.setIcon("layui-icon "+menu.getIcon());
         return (menuDao.update(menu) > 0) ? Result.ok().message("修改成功") : Result.error().message("修改失败");
-
     }
 
     @Override
     public Result<MyMenu> save(MyMenu menu) {
+        menu.setIcon("layui-icon "+menu.getIcon());
         return (menuDao.save(menu) > 0) ? Result.ok().message("添加成功") : Result.error().message("添加失败");
-
     }
 
-    //如果这里删除了菜单树的父节点，把它的子节点一并删除
+    /**
+     * 如果这里删除了菜单树的父节点，把它的子节点一并删除
+     * @param id
+     * @return
+     */
     @Override
     public Result delete(Integer id) {
-        menuDao.deleteById(id);
-        menuDao.deleteByParentId(id);
-        return Result.ok().message("删除成功");
+        Integer count = roleMenuDao.countRoleMenuByRoleId(id);
+        if (count == 0){
+            menuDao.deleteById(id);
+            List<Integer> list = menuDao.selectByParentId(id);
+            if(list.size()>0){
+                for (Integer parentId: list){
+                    menuDao.deleteByParentId(parentId);
+                }
+                menuDao.deleteByParentId(id);
+            }
+            return Result.ok().message("删除成功");
+        }
+        else {
+            return Result.error().message("已经有角色分配该菜单，无法删除");
+        }
+
     }
 
     @Override
